+++ tests/main_tests.rs
@@ -179,5 +179,8 @@ mod test_count_var;
 #[path = "../tests_all/test_count_view_flexibility.rs"]
 mod test_count_view_flexibility;
 
+#[path = "../tests_all/test_bool_xor.rs"]
+mod test_bool_xor;
+
 #[path = "../tests_all/test_element_computed_index.rs"]
 mod test_element_computed_index;